feat: implement an example about method overloading and type casting
In this example we see what will happen if run this code, having all methods active. In comparison, we will see that Java sees 100 as an integer literal. If we comment-out print(int a) Java will handle 100 as the next safely implicit cast, which is long and double respective. When commenting out every save implicit cast, and we try to print 100 Java won't compile. So we have to explicitly cast 100 to short.